requirements.txt
coincurve==20.0.0
ecdsa==0.18.0
pycryptodome==3.14.1
pysha3==1.0.2
